Add files via upload
This program can be run in python and imports the data recorded by the World Bank, the Women, Business and Law Index (WBL Index) from 1971 to 2022, as well as ISO country and region codes. The program reads in the data and converts it into the indexdf dataframe format. Then it creates an animated choropleth map that shows how the WBL Index changes in countries over time. The python program opens a new browser window and the user is able to “play” through the data by clicking the play button at the bottom near the year legend.